scion-stack: fix error handling for ScionAsyncUdpSocket to ensure proper quinn timeouts
After suspend, there is a race condition between the inner snaptun and the edgetun quic connection. If the timeout arrives first in the inner quic connection, the connection driver of the edgetun connection quits resulting in a stale connection. If the `ScionAsyncUdpSocket` doesn't return the error, we prevent the edgetun connection driver to quit which will result in a timeout to properly close the connection and entering the daemon re-connect cycle.

GitOrigin-RevId: 9c83d31b8f7cd4f9a27aa648d72029a978974d71